added vehicle controller

+++ backend/WPRProject/Tables/SubscriptionCoverage.cs
@@ -5,6 +5,6 @@ namespace WPRProject.Tables
     public class SubscriptionCoverage
     {
         [Key] public int Id { get; set; }
-        public int DailyRentCoverage {  get; set; }
+        public double DailyRentCoverage {  get; set; }
     }
 }
